requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.65.0
streamlit>=1.32.0 